# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "1.0.0"
  description: VoxAgent AI Voice Bot Platform API
servers:
  - url: /api
paths:
...
  /v1/compliance/dnc:
    get:
      operationId: listDncEntries
      summary: List tenant do-not-call entries
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: DNC entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DncEntry"
    post:
      operationId: createDncEntry
      summary: Add a tenant do-not-call entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DncEntryInput"
      responses:
        "201":
          description: DNC entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DncEntry"
  /v1/compliance/dnc/{id}:
    delete:
      operationId: deleteDncEntry
      summary: Remove a tenant do-not-call entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: DNC entry removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DncEntry"
  /v1/compliance/dnc/import:
    post:
      operationId: importDncCsv
      summary: Import DNC phone numbers from CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DncCsvImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsvImportResult"
...
  /v1/compliance/consents:
    get:
      operationId: listConsentLedger
      summary: List tenant consent events
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Consent events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsentEntry"
    post:
      operationId: recordConsent
      summary: Record a grant or revocation consent event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentInput"
      responses:
        "201":
          description: Consent event recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentEntry"
  /v1/compliance/consents/{id}/revoke:
    post:
      operationId: revokeConsent
      summary: Revoke a consent ledger entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Revocation event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentEntry"
...
  /v1/compliance/profiles/{jurisdictionCode}:
    put:
      operationId: updateComplianceProfile
      summary: Create or update a jurisdiction profile
      parameters:
        - in: path
          name: jurisdictionCode
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceProfile"
...
  /v1/compliance/opt-out:
    post:
      operationId: recordInCallOptOut
      summary: Add a caller to DNC during a live call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptOutInput"
      responses:
        "201":
          description: Opt-out recorded
          content:
            application/json:
              schema:
                type: object

  # ─── Campaigns ──────────────────────────────────────────────────
...
components:
...
  schemas:
...
    DncEntry:
      type: object
      required: [id, tenantId, phoneNumber, source, addedAt, createdAt]
      properties:
        id: { type: string }
        tenantId: { type: string }
        phoneNumber: { type: string }
        source: { type: string }
        reason: { type: ["string", "null"] }
        addedAt: { type: string, format: date-time }
        expiresAt: { type: ["string", "null"], format: date-time }
        createdByUserId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    DncEntryInput:
      type: object
      required: [phoneNumber]
      properties:
        phoneNumber: { type: string }
        source: { type: string }
        reason: { type: string }
        expiresAt: { type: ["string", "null"], format: date-time }
    DncCsvImportInput:
      type: object
      required: [csv]
      properties:
        csv: { type: string }
    ConsentEntry:
      type: object
      required: [id, tenantId, phoneNumber, consentType, status, source, capturedAt, createdAt]
      properties:
        id: { type: string }
        tenantId: { type: string }
        phoneNumber: { type: string }
        consentType: { type: string, enum: [VOICE_CALLING, RECORDING] }
        status: { type: string, enum: [GRANTED, REVOKED] }
        source: { type: string }
        capturedAt: { type: string, format: date-time }
        expiresAt: { type: ["string", "null"], format: date-time }
        evidence: { type: ["string", "null"] }
        actorUserId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ConsentInput:
      type: object
      required: [phoneNumber, consentType, status]
      properties:
        phoneNumber: { type: string }
        consentType: { type: string, enum: [VOICE_CALLING, RECORDING] }
        status: { type: string, enum: [GRANTED, REVOKED] }
        source: { type: string }
        evidence: { type: string }
        expiresAt: { type: ["string", "null"], format: date-time }
    ComplianceProfile:
      type: object
      required: [id, tenantId, jurisdictionCode, displayName, enabled, timezone, callingWindowStart, callingWindowEnd, allowedDays, holidays, requireConsent, requireRecordingConsent, blockOnHoliday, createdAt, updatedAt]
      properties:
        id: { type: string }
        tenantId: { type: string }
        jurisdictionCode: { type: string }
        displayName: { type: string }
        enabled: { type: boolean }
        timezone: { type: string }
        callingWindowStart: { type: string }
        callingWindowEnd: { type: string }
        allowedDays:
          type: array
          items: { type: string }
        holidays:
          type: array
          items: { type: string }
        requireConsent: { type: boolean }
        requireRecordingConsent: { type: boolean }
        mandatoryDisclosureText: { type: ["string", "null"] }
        blockOnHoliday: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ComplianceProfileInput:
      type: object
      required: [displayName, timezone, callingWindowStart, callingWindowEnd, allowedDays, requireConsent, requireRecordingConsent, blockOnHoliday]
      properties:
        displayName: { type: string }
        enabled: { type: boolean }
        timezone: { type: string }
        callingWindowStart: { type: string }
        callingWindowEnd: { type: string }
        allowedDays:
          type: array
          items: { type: string }
        holidays:
          type: array
          items: { type: string }
        requireConsent: { type: boolean }
        requireRecordingConsent: { type: boolean }
        mandatoryDisclosureText: { type: ["string", "null"] }
        blockOnHoliday: { type: boolean }
...
    CsvImportResult:
      type: object
      required: [imported, rejected, errors]
      properties:
        imported: { type: integer }
        rejected: { type: integer }
        errors:
          type: array
          items: { type: string }
    OptOutInput:
      type: object
      required: [phoneNumber]
      properties:
        phoneNumber: { type: string }
        callId: { type: string }
        reason: { type: string }
...